requirements.txt
asgiref
Django
django-ckeditor
django-crispy-forms
django-js-asset
gunicorn
Pillow
pytz
sqlparse
whitenoise
